Fix startup failure due to timing issues with SleuthAnnotationConfiguration
The SleuthAnnotationConfiguration uses `@PostConstruct` to initialize an
advisor, a pointcut and a bean factory (based on the advisor type). However, it
seems that some Spring Boot 3 autoconfiguration asserts if the pointcut is not
null, before the `@PostConstruct` code is performed, which causes the assertion
to always return false.

To fix this, I moved initializing code into the getters of the pointcut and
advice attributes. However, this is by no means an educated solution and more
result of guess work and trial and error.