Resolve cast attempt from `ImmutableAssertingSpan` to `brave.Span`
`ImmutableAssertingSpan` causes cast issues during runtime. More precisely, when
`@NewSpan` is processed a `Tracer` is used to create a new span. The concrete
type of the span depends on the tracer implementation added as a dependency to
the project. In this case it is a `brave.Span`. This concrete type can not be
cast to a `ImmutableAssertingSpan`, because `AssertingSpan` provides methods
that overlap with the concrete implementation.

As a solution I removed the `wrap` call that attempts to change the returned
span from the tracer into an `AssertingSpan`. Surprisingly, this does not affect
the tagging and event creation process, which also uses `wrap`.