Eliminate dependency on Sleuth's docs package for annotation port
The Sleuth code uses the docs `AssertedSpan` interface to add tags and
events to spans. However, since the `AssertedSpan` interface is nothing more
than a wrapper around a normal span, it and its related classes/interfaces can
be removed and be replaced with simple `tags()` and `event()` calls on the
original span.